server/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api>=2.0.2",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.1",
]
